+++ packages/examples/fj_playground/python-modules/lark/utils.py
@@ -1,9 +1,8 @@
 import unicodedata
 import os
 from functools import reduce
-from itertools import product
 from collections import deque
-from typing import Callable, Iterator, List, Optional, Tuple, Type, TypeVar, Union, Dict, Any, Sequence, Iterable
+from typing import Callable, Iterator, List, Optional, Tuple, Type, TypeVar, Union, Dict, Any, Sequence
 
 ###{standalone
 import sys, re
@@ -21,14 +20,14 @@ NO_VALUE = object()
 T = TypeVar("T")
 
 
-def classify(seq: Iterable, key: Optional[Callable] = None, value: Optional[Callable] = None) -> Dict:
+def classify(seq: Sequence, key: Optional[Callable] = None, value: Optional[Callable] = None) -> Dict:
     d: Dict[Any, Any] = {}
     for item in seq:
         k = key(item) if (key is not None) else item
         v = value(item) if (value is not None) else item
-        try:
+        if k in d:
             d[k].append(v)
-        except KeyError:
+        else:
             d[k] = [v]
     return d
 
@@ -189,7 +188,7 @@ def dedup_list(l: List[T]) -> List[T]:
     dedup = set()
     # This returns None, but that's expected
     return [x for x in l if not (x in dedup or dedup.add(x))]  # type: ignore[func-returns-value]
-    # 2x faster (ordered in PyPy and CPython 3.6+, guaranteed to be ordered in Python 3.7+)
+    # 2x faster (ordered in PyPy and CPython 3.6+, gaurenteed to be ordered in Python 3.7+)
     # return list(dict.fromkeys(l))
 
 
@@ -214,7 +213,7 @@ class Enumerator(Serialize):
 
 def combine_alternatives(lists):
     """
-    Accepts a list of alternatives, and enumerates all their possible concatenations.
+    Accepts a list of alternatives, and enumerates all their possible concatinations.
 
     Examples:
         >>> combine_alternatives([range(2), [4,5]])
@@ -229,7 +228,9 @@ def combine_alternatives(lists):
     if not lists:
         return [[]]
     assert all(l for l in lists), lists
-    return list(product(*lists))
+    init = [[x] for x in lists[0]]
+    return reduce(lambda a,b: [i+[j] for i in a for j in b], lists[1:], init)
+
 
 try:
     import atomicwrites
@@ -267,8 +268,15 @@ class fzset(frozenset):
 
 
 def classify_bool(seq: Sequence, pred: Callable) -> Any:
+    true_elems = []
     false_elems = []
-    true_elems = [elem for elem in seq if pred(elem) or false_elems.append(elem)]  # type: ignore[func-returns-value]
+
+    for elem in seq:
+        if pred(elem):
+            true_elems.append(elem)
+        else:
+            false_elems.append(elem)
+
     return true_elems, false_elems