+++ packages/python_ffi_cpython_dart/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+- Fixed an issue with loading python modules.
 - Updated dependencies.
 
 ## 0.2.2